// src/shared/types/property.types.ts
export interface Property {
  id: string;
  title: string;
  description: string;
  location: string;
  salePrice: number;
  bedrooms: number;
  bathrooms: number;
  sqft: number;
  parking: number;
  thumbnailUrl: string;
  pictureUrl: string;
  dateListed: string;
  yearBuilt: number;
  saved: boolean;
}